+++ docs/changes.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+1.3.0 (2018/??)
+---------------
+
+Adds support for following cloud storage:
+
+* Google Cloud Storage
+
 1.2.0 (2018/10)
 ---------------